Découper le panneau d'actions des mesures en quatre formulaires
Une fonction anonyme de 331 lignes dispatchait rejet, dépublication, retrait et
brouillon par une chaîne de if. Chaque formulaire déclare désormais la variante
d'action qu'il traite, donc l'union reste discriminée.
